fix: 평균 별점 오류 수정 및 타입 변환

+++ src/main/java/com/example/tripKo/domain/food/dto/response/FoodDetailsResponse.java
@@ -26,7 +26,7 @@ public class FoodDetailsResponse {
         private String name;
         private String location;
         private String image;
-        private float averageRating;
+        private double averageRating;
     }
 
     public FoodDetailsResponse(Food food) {
