+++ src/main/java/com/example/tripKo/domain/place/dto/response/search/PlaceRestaurantResponse.java
@@ -14,7 +14,7 @@ public class PlaceRestaurantResponse {
   private String summary;
   private String image;
   private String address;
-  private float averageRating;
+  private double averageRating;
   private Boolean isWished;
 
   @Builder
